chore: move PackageBuildProvenance into a separate file
